# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pocket Number API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List users (paginated, newest-first)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: before
          required: false
          schema:
            type: integer
          description: Cursor — return users older than this user ID
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminUserItem"
                  hasMore:
                    type: boolean
                  nextCursor:
                    type: integer
                    nullable: true
                required:
                  - users
                  - hasMore
                  - nextCursor
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — caller is not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Get user details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — caller is not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}/suspend:
    patch:
      operationId: adminSuspendUser
      tags: [admin]
      summary: Suspend a user account (blocks future logins)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserItem"
        "400":
          description: Invalid request (e.g. attempting to suspend yourself)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — caller is not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}/restore:
    patch:
      operationId: adminRestoreUser
      tags: [admin]
      summary: Restore (unsuspend) a user account
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — caller is not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Devices (notification foundation — no push provider yet)
  # ---------------------------------------------------------------------------
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      description: Creates a new account and sends an OTP to the email for verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registration successful, OTP sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  devOtp:
                    type: string
                    description: OTP code for testing — present only in development mode, never in production
                required:
                  - message
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpInput"
      responses:
        "200":
          description: OTP verified successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/resend-otp:
    post:
      operationId: resendOtp
      tags: [auth]
      summary: Resend OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendOtpInput"
      responses:
        "200":
          description: OTP resent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  devOtp:
                    type: string
                    description: OTP code for testing — present only in development mode, never in production
                required:
                  - message
        "400":
          description: Email not found or already verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Email not verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout and mark user offline
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        pocketNumber:
          type: string
          description: "9-digit unique number with prefix 71/73/77/700 (e.g. 713456789)"
        name:
          type: string
        email:
          type: string
        isVerified:
          type: boolean
        isOnline:
          type: boolean
          description: Whether the user is currently online
        lastSeenAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of the user's last activity
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - pocketNumber
        - name
        - email
        - isVerified
        - isOnline
        - lastSeenAt
        - createdAt
...
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    RegisterInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
        email:
          type: string
        password:
          type: string
          minLength: 8
          maxLength: 100
      required:
        - name
        - email
        - password
    VerifyOtpInput:
      type: object
      properties:
        email:
          type: string
        code:
          type: string
          minLength: 6
          maxLength: 6
      required:
        - email
        - code
    ResendOtpInput:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 1
      required:
        - email
        - password
...
    AdminUserItem:
      type: object
      description: A user record as seen by an admin
      properties:
        id:
          type: integer
        pocketNumber:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        isVerified:
          type: boolean
        isOnline:
          type: boolean
        isSuspended:
          type: boolean
        lastSeenAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - pocketNumber
        - name
        - email
        - role
        - isVerified
        - isOnline
        - isSuspended
        - lastSeenAt
        - createdAt

    # -------------------------------------------------------------------------
    # Messages
    # -------------------------------------------------------------------------
...