# lib/api-zod/src/generated/types/adminUserItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Pocket Number API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUserItemRole } from './adminUserItemRole';

/**
 * A user record as seen by an admin
 */
export interface AdminUserItem {
  id: number;
  pocketNumber: string;
  name: string;
  email: string;
  role: AdminUserItemRole;
  isVerified: boolean;
  isOnline: boolean;
  isSuspended: boolean;
  lastSeenAt: Date | null;
  createdAt: Date;
}